Imported modifications from sux4j-test
