+++ jgrapht-unimi-dsi/src/main/java/org/jgrapht/sux4j/SuccinctDirectedGraph.java
@@ -19,32 +19,21 @@
 package org.jgrapht.sux4j;
 
 import java.io.Serializable;
-import java.util.Arrays;
 import java.util.Iterator;
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.Set;
-import java.util.function.Function;
-import java.util.function.Supplier;
 
 import org.jgrapht.Graph;
 import org.jgrapht.GraphIterables;
-import org.jgrapht.GraphType;
-import org.jgrapht.Graphs;
 import org.jgrapht.alg.util.Pair;
-import org.jgrapht.graph.AbstractGraph;
-import org.jgrapht.graph.DefaultGraphType;
 import org.jgrapht.opt.graph.sparse.SparseIntDirectedGraph;
 
 import com.google.common.collect.Iterables;
 
-import it.unimi.dsi.fastutil.ints.IntArrays;
 import it.unimi.dsi.fastutil.ints.IntIntPair;
-import it.unimi.dsi.fastutil.ints.IntSets;
 import it.unimi.dsi.fastutil.longs.LongBigListIterator;
-import it.unimi.dsi.fastutil.longs.LongIterator;
 import it.unimi.dsi.fastutil.objects.ObjectOpenHashSet;
-import it.unimi.dsi.fastutil.objects.ObjectSets;
 import it.unimi.dsi.sux4j.util.EliasFanoIndexedMonotoneLongBigList;
 import it.unimi.dsi.sux4j.util.EliasFanoIndexedMonotoneLongBigList.EliasFanoIndexedMonotoneLongBigListIterator;
 import it.unimi.dsi.sux4j.util.EliasFanoMonotoneLongBigList;
@@ -80,118 +69,12 @@ import it.unimi.dsi.sux4j.util.EliasFanoMonotoneLongBigList;
 
 public class SuccinctDirectedGraph
     extends
-    AbstractGraph<Integer, IntIntPair>
+    AbstractSuccinctDirectedGraph<IntIntPair>
     implements
     Serializable
 {
     private static final long serialVersionUID = 0L;
-    protected static final String UNMODIFIABLE = "this graph is unmodifiable";
 
-    /**
-     * Turns all lists of successors into a single monotone sequence by representing an arc
-     * <var>x</var>&nbsp;&rarr;&nbsp;<var>y</var> as <var>x</var><var>n</var> + <var>y</var>.
-     *
-     * @param <E> the graph edge type
-     */
-    private final static class CumulativeSuccessors<E>
-        implements
-        LongIterator
-    {
-        private final Graph<Integer, E> graph;
-        private final long n;
-        private final Function<Integer, Iterable<E>> succ;
-        private final boolean strict;
-
-        private int x = -1, d, i, e;
-        private long next = -1;
-        private int[] s = IntArrays.EMPTY_ARRAY;
-
-        public CumulativeSuccessors(
-            final Graph<Integer, E> graph, final Function<Integer, Iterable<E>> succ,
-            final boolean strict)
-        {
-            this.n = graph.iterables().vertexCount();
-            this.graph = graph;
-            this.succ = succ;
-            this.strict = strict;
-        }
-
-        @Override
-        public boolean hasNext()
-        {
-            if (next != -1)
-                return true;
-            if (x == n)
-                return false;
-            while (i == d) {
-                if (++x == n)
-                    return false;
-                int d = 0;
-                for (final E e : succ.apply(x)) {
-                    s = IntArrays.grow(s, d + 1);
-                    s[d++] = Graphs.getOppositeVertex(graph, e, x);
-                }
-                Arrays.sort(s, 0, d);
-                this.d = d;
-                i = 0;
-            }
-            // The predecessor list will not be indexed, so we can gain a few bits of space by
-            // subtracting the edge position in the list
-            next = s[i] + x * n - (strict ? 0 : e++);
-            i++;
-            return true;
-        }
-
-        @Override
-        public long nextLong()
-        {
-            if (!hasNext())
-                throw new NoSuchElementException();
-            final long result = next;
-            next = -1;
-            return result;
-        }
-    }
-
-    /**
-     * Iterates over the cumulative degrees (starts with a zero).
-     */
-    private final static class CumulativeDegrees
-        implements
-        LongIterator
-    {
-        private final Function<Integer, Integer> degreeOf;
-        private final int n;
-        private int i = -1;
-        private long cumul = 0;
-
-        public CumulativeDegrees(final int n, final Function<Integer, Integer> degreeOf)
-        {
-            this.n = n;
-            this.degreeOf = degreeOf;
-        }
-
-        @Override
-        public boolean hasNext()
-        {
-            return i < n;
-        }
-
-        @Override
-        public long nextLong()
-        {
-            if (!hasNext())
-                throw new NoSuchElementException();
-            if (i == -1)
-                return ++i;
-            return cumul += degreeOf.apply(i++);
-        }
-    }
-
-    /** The number of vertices in the graph. */
-    private final int n;
-    /** The number of edges in the graph. */
-    private final int m;
     /** The cumulative list of outdegrees. */
     private final EliasFanoIndexedMonotoneLongBigList cumulativeOutdegrees;
     /** The cumulative list of indegrees. */
@@ -210,6 +93,7 @@ public class SuccinctDirectedGraph
      */
     public <E> SuccinctDirectedGraph(final Graph<Integer, E> graph)
     {
+        super((int) graph.iterables().vertexCount(), (int) graph.iterables().edgeCount());
 
         if (graph.getType().isUndirected())
             throw new IllegalArgumentException("This class supports directed graphs only");
@@ -224,9 +108,6 @@ public class SuccinctDirectedGraph
                 "The number of edges (" + iterables.edgeCount() + ") is greater than "
                     + Integer.MAX_VALUE);
 
-        n = (int) iterables.vertexCount();
-        m = (int) iterables.edgeCount();
-
         cumulativeOutdegrees = new EliasFanoIndexedMonotoneLongBigList(
             n + 1, m, new CumulativeDegrees(n, graph::outDegreeOf));
         cumulativeIndegrees =
@@ -235,7 +116,8 @@ public class SuccinctDirectedGraph
         assert cumulativeIndegrees.getLong(cumulativeIndegrees.size64() - 1) == m;
 
         successors = new EliasFanoIndexedMonotoneLongBigList(
-            m, (long) n * n, new CumulativeSuccessors<>(graph, iterables::outgoingEdgesOf, true));
+            m, (long) n << sourceShift,
+            new CumulativeSuccessors<>(graph, iterables::outgoingEdgesOf, true));
         predecessors = new EliasFanoIndexedMonotoneLongBigList(
             m, (long) n * n - m,
             new CumulativeSuccessors<>(graph, iterables::incomingEdgesOf, false));
@@ -259,64 +141,15 @@ public class SuccinctDirectedGraph
     }
 
     @Override
-    public Supplier<Integer> getVertexSupplier()
-    {
-        return null;
-    }
-
-    @Override
-    public Supplier<IntIntPair> getEdgeSupplier()
-    {
-        return null;
-    }
-
-    @Override
-    public IntIntPair addEdge(final Integer sourceVertex, final Integer targetVertex)
-    {
-        throw new UnsupportedOperationException(UNMODIFIABLE);
-    }
-
-    @Override
-    public boolean addEdge(
-        final Integer sourceVertex, final Integer targetVertex, final IntIntPair e)
-    {
-        throw new UnsupportedOperationException(UNMODIFIABLE);
-    }
-
-    @Override
-    public Integer addVertex()
-    {
-        throw new UnsupportedOperationException(UNMODIFIABLE);
-    }
-
-    @Override
-    public boolean addVertex(final Integer v)
-    {
-        throw new UnsupportedOperationException(UNMODIFIABLE);
-    }
-
-    @Override
     public boolean containsEdge(final IntIntPair e)
     {
-        return successors.indexOfUnsafe(e.firstInt() * (long) n + e.secondInt()) != -1;
-    }
-
-    @Override
-    public boolean containsVertex(final Integer v)
-    {
-        return v >= 0 && v < n;
+        return successors.indexOfUnsafe(((long) e.firstInt() << sourceShift) + e.secondInt()) != -1;
     }
 
     @Override
     public Set<IntIntPair> edgeSet()
     {
         return new ObjectOpenHashSet<>(iterables().edges().iterator());
-    }
-
-    @Override
-    public int degreeOf(final Integer vertex)
-    {
-        return inDegreeOf(vertex) + outDegreeOf(vertex);
     }
 
     @Override
@@ -349,7 +182,7 @@ public class SuccinctDirectedGraph
 
         for (int i = d; i-- != 0;) {
             final long source = iterator.nextLong() - base--;
-            s.add(IntIntPair.of((int) source, target));
+            s.add(IntIntPair.of((int) source, t));
         }
 
         return s;
@@ -371,33 +204,9 @@ public class SuccinctDirectedGraph
         final ObjectOpenHashSet<IntIntPair> s = new ObjectOpenHashSet<>();
         for (int e = (int) result[0]; e < (int) result[1]; e++) {
             final long t = successors.getLong(e);
-            s.add(IntIntPair.of((int) (t / n), (int) (t % n)));
+            s.add(IntIntPair.of((int) (t >>> sourceShift), (int) (t & targetMask)));
         }
         return s;
-    }
-
-    @Override
-    public IntIntPair removeEdge(final Integer sourceVertex, final Integer targetVertex)
-    {
-        throw new UnsupportedOperationException(UNMODIFIABLE);
-    }
-
-    @Override
-    public boolean removeEdge(final IntIntPair e)
-    {
-        throw new UnsupportedOperationException(UNMODIFIABLE);
-    }
-
-    @Override
-    public boolean removeVertex(final Integer v)
-    {
-        throw new UnsupportedOperationException(UNMODIFIABLE);
-    }
-
-    @Override
-    public Set<Integer> vertexSet()
-    {
-        return IntSets.fromTo(0, n);
     }
 
     @Override
@@ -412,74 +221,49 @@ public class SuccinctDirectedGraph
         return e.secondInt();
     }
 
-    @Override
-    public GraphType getType()
+    /**
+     * Returns the index associated with the given edge.
+     *
+     * @param e an edge of the graph.
+     * @return the index associated with the edge, or &minus;1 if the edge is not part of the graph.
+     * @see #getEdgeFromIndex(int)
+     */
+    public int getIndexFromEdge(final IntIntPair e)
     {
-        return new DefaultGraphType.Builder()
-            .directed().weighted(false).modifiable(false).allowMultipleEdges(false)
-            .allowSelfLoops(true).build();
+        final int source = e.firstInt();
+        final int target = e.secondInt();
+        if (source < 0 || source >= n || target < 0 || target >= n)
+            throw new IllegalArgumentException();
+        return (int) successors.indexOfUnsafe(((long) source << sourceShift) + target);
     }
 
-    @Override
-    public double getEdgeWeight(final IntIntPair e)
+    /**
+     * Returns the edge with given index.
+     *
+     * @param i an index between 0 (included) and the number of edges (excluded).
+     * @return the pair with index {@code i}.
+     * @see #getIndexFromEdge(IntIntPair)
+     */
+    public IntIntPair getEdgeFromIndex(final int i)
     {
-        return 1.0;
-    }
-
-    @Override
-    public void setEdgeWeight(final IntIntPair e, final double weight)
-    {
-        throw new UnsupportedOperationException(UNMODIFIABLE);
+        if (i < 0 || i >= m)
+            throw new IllegalArgumentException();
+        final long t = successors.getLong(i);
+        return IntIntPair.of((int) (t >>> sourceShift), (int) (t & targetMask));
     }
 
     @Override
     public IntIntPair getEdge(final Integer sourceVertex, final Integer targetVertex)
     {
-        final long index = successors.indexOfUnsafe(sourceVertex * (long) n + targetVertex);
+        final long index =
+            successors.indexOfUnsafe(((long) sourceVertex << sourceShift) + targetVertex);
         return index != -1 ? IntIntPair.of(sourceVertex, targetVertex) : null;
     }
 
     @Override
     public boolean containsEdge(final Integer sourceVertex, final Integer targetVertex)
     {
-        return successors.indexOfUnsafe(sourceVertex * (long) n + targetVertex) != -1;
-    }
-
-    @Override
-    public Set<IntIntPair> getAllEdges(final Integer sourceVertex, final Integer targetVertex)
-    {
-        final IntIntPair edge = getEdge(sourceVertex, targetVertex);
-        return edge == null ? ObjectSets.emptySet() : ObjectSets.singleton(edge);
-    }
-
-    /**
-     * Ensures that the specified vertex exists in this graph, or else throws exception.
-     *
-     * @param v vertex
-     * @return <code>true</code> if this assertion holds.
-     * @throws IllegalArgumentException if specified vertex does not exist in this graph.
-     */
-    @Override
-    protected boolean assertVertexExist(final Integer v)
-    {
-        if (v < 0 || v >= n)
-            throw new IllegalArgumentException();
-        return true;
-    }
-
-    /**
-     * Ensures that the specified edge exists in this graph, or else throws exception.
-     *
-     * @param e edge
-     * @return <code>true</code> if this assertion holds.
-     * @throws IllegalArgumentException if specified edge does not exist in this graph.
-     */
-    protected boolean assertEdgeExist(final Integer e)
-    {
-        if (e < 0 || e >= m)
-            throw new IllegalArgumentException();
-        return true;
-
+        return successors.indexOfUnsafe(((long) sourceVertex << sourceShift) + targetVertex) != -1;
     }
 
     private final static class SuccinctGraphIterables
@@ -521,6 +305,9 @@ public class SuccinctDirectedGraph
         @Override
         public Iterable<IntIntPair> edges()
         {
+            final int sourceShift = graph.sourceShift;
+            final long targetMask = graph.targetMask;
+
             return () -> new Iterator<>()
             {
                 private final EliasFanoIndexedMonotoneLongBigListIterator iterator = graph.successors.iterator();
@@ -536,7 +323,7 @@ public class SuccinctDirectedGraph
                 public IntIntPair next()
                 {
                     final long t = iterator.nextLong();
-                    return IntIntPair.of((int) (t / n), (int) (t % n));
+                    return IntIntPair.of((int) (t >>> sourceShift), (int) (t & targetMask));
                 }
 
             };
@@ -592,6 +379,9 @@ public class SuccinctDirectedGraph
         @Override
         public Iterable<IntIntPair> outgoingEdgesOf(final Integer vertex)
         {
+            final int sourceShift = graph.sourceShift;
+            final long targetMask = graph.targetMask;
+
             graph.assertVertexExist(vertex);
             final long[] result = new long[2];
             graph.cumulativeOutdegrees.get(vertex, result);
@@ -613,7 +403,7 @@ public class SuccinctDirectedGraph
                     if (!hasNext())
                         throw new NoSuchElementException();
                     final long t = successors.getLong(e++);
-                    return IntIntPair.of((int) (t / n), (int) (t % n));
+                    return IntIntPair.of((int) (t >>> sourceShift), (int) (t & targetMask));
                 }
 
             };
