+++ jgrapht-unimi-dsi/src/main/java/org/jgrapht/sux4j/SuccinctIntDirectedGraph.java
@@ -19,31 +19,22 @@
 package org.jgrapht.sux4j;
 
 import java.io.Serializable;
-import java.util.Arrays;
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.Set;
-import java.util.function.Function;
-import java.util.function.Supplier;
 
 import org.jgrapht.Graph;
 import org.jgrapht.GraphIterables;
-import org.jgrapht.GraphType;
-import org.jgrapht.Graphs;
 import org.jgrapht.alg.util.Pair;
-import org.jgrapht.graph.AbstractGraph;
-import org.jgrapht.graph.DefaultGraphType;
 import org.jgrapht.opt.graph.sparse.SparseIntDirectedGraph;
 
 import com.google.common.collect.Iterables;
 
-import it.unimi.dsi.fastutil.ints.IntArrays;
 import it.unimi.dsi.fastutil.ints.IntIterator;
 import it.unimi.dsi.fastutil.ints.IntOpenHashSet;
 import it.unimi.dsi.fastutil.ints.IntSet;
 import it.unimi.dsi.fastutil.ints.IntSets;
 import it.unimi.dsi.fastutil.longs.LongBigListIterator;
-import it.unimi.dsi.fastutil.longs.LongIterator;
 import it.unimi.dsi.sux4j.util.EliasFanoIndexedMonotoneLongBigList;
 import it.unimi.dsi.sux4j.util.EliasFanoMonotoneLongBigList;
 
@@ -78,118 +69,11 @@ import it.unimi.dsi.sux4j.util.EliasFanoMonotoneLongBigList;
 
 public class SuccinctIntDirectedGraph
     extends
-    AbstractGraph<Integer, Integer>
+    AbstractSuccinctDirectedGraph<Integer>
     implements
     Serializable
 {
     private static final long serialVersionUID = 0L;
-    protected static final String UNMODIFIABLE = "this graph is unmodifiable";
-
-    /**
-     * Turns all lists of successors into a single monotone sequence by representing an arc
-     * <var>x</var>&nbsp;&rarr;&nbsp;<var>y</var> as <var>x</var><var>n</var> + <var>y</var>.
-     *
-     * @param <E> the graph edge type
-     */
-    private final static class CumulativeSuccessors<E>
-        implements
-        LongIterator
-    {
-        private final Graph<Integer, E> graph;
-        private final long n;
-        private final Function<Integer, Iterable<E>> succ;
-        private final boolean strict;
-
-        private int x = -1, d, i, e;
-        private long next = -1;
-        private int[] s = IntArrays.EMPTY_ARRAY;
-
-        public CumulativeSuccessors(
-            final Graph<Integer, E> graph, final Function<Integer, Iterable<E>> succ,
-            final boolean strict)
-        {
-            this.n = graph.iterables().vertexCount();
-            this.graph = graph;
-            this.succ = succ;
-            this.strict = strict;
-        }
-
-        @Override
-        public boolean hasNext()
-        {
-            if (next != -1)
-                return true;
-            if (x == n)
-                return false;
-            while (i == d) {
-                if (++x == n)
-                    return false;
-                int d = 0;
-                for (final E e : succ.apply(x)) {
-                    s = IntArrays.grow(s, d + 1);
-                    s[d++] = Graphs.getOppositeVertex(graph, e, x);
-                }
-                Arrays.sort(s, 0, d);
-                this.d = d;
-                i = 0;
-            }
-            // The predecessor list will not be indexed, so we can gain a few bits of space by
-            // subtracting the edge position in the list
-            next = s[i] + x * n - (strict ? 0 : e++);
-            i++;
-            return true;
-        }
-
-        @Override
-        public long nextLong()
-        {
-            if (!hasNext())
-                throw new NoSuchElementException();
-            final long result = next;
-            next = -1;
-            return result;
-        }
-    }
-
-    /**
-     * Iterates over the cumulative degrees (starts with a zero).
-     */
-    private final static class CumulativeDegrees
-        implements
-        LongIterator
-    {
-        private final Function<Integer, Integer> degreeOf;
-        private final int n;
-        private int i = -1;
-        private long cumul = 0;
-
-        public CumulativeDegrees(final int n, final Function<Integer, Integer> degreeOf)
-        {
-            this.n = n;
-            this.degreeOf = degreeOf;
-        }
-
-        @Override
-        public boolean hasNext()
-        {
-            return i < n;
-        }
-
-        @Override
-        public long nextLong()
-        {
-            if (!hasNext())
-                throw new NoSuchElementException();
-            if (i == -1)
-                return ++i;
-            return cumul += degreeOf.apply(i++);
-        }
-    }
-
-    /** The number of vertices in the graph. */
-    private final int n;
-    /** The number of edges in the graph. */
-    private final int m;
     /** The cumulative list of outdegrees. */
     private final EliasFanoIndexedMonotoneLongBigList cumulativeOutdegrees;
     /** The cumulative list of indegrees. */
@@ -208,6 +92,7 @@ public class SuccinctIntDirectedGraph
      */
     public <E> SuccinctIntDirectedGraph(final Graph<Integer, E> graph)
     {
+        super((int) graph.iterables().vertexCount(), (int) graph.iterables().edgeCount());
 
         if (graph.getType().isUndirected())
             throw new IllegalArgumentException("This class supports directed graphs only");
@@ -222,9 +107,6 @@ public class SuccinctIntDirectedGraph
                 "The number of edges (" + iterables.edgeCount() + ") is greater than "
                     + Integer.MAX_VALUE);
 
-        n = (int) iterables.vertexCount();
-        m = (int) iterables.edgeCount();
-
         cumulativeOutdegrees = new EliasFanoIndexedMonotoneLongBigList(
             n + 1, m, new CumulativeDegrees(n, graph::outDegreeOf));
         cumulativeIndegrees =
@@ -233,7 +115,8 @@ public class SuccinctIntDirectedGraph
         assert cumulativeIndegrees.getLong(cumulativeIndegrees.size64() - 1) == m;
 
         successors = new EliasFanoIndexedMonotoneLongBigList(
-            m, (long) n * n, new CumulativeSuccessors<>(graph, iterables::outgoingEdgesOf, true));
+            m, (long) n << sourceShift,
+            new CumulativeSuccessors<>(graph, iterables::outgoingEdgesOf, true));
         predecessors = new EliasFanoIndexedMonotoneLongBigList(
             m, (long) n * n - m,
             new CumulativeSuccessors<>(graph, iterables::incomingEdgesOf, false));
@@ -257,63 +140,15 @@ public class SuccinctIntDirectedGraph
     }
 
     @Override
-    public Supplier<Integer> getVertexSupplier()
-    {
-        return null;
-    }
-
-    @Override
-    public Supplier<Integer> getEdgeSupplier()
-    {
-        return null;
-    }
-
-    @Override
-    public Integer addEdge(final Integer sourceVertex, final Integer targetVertex)
-    {
-        throw new UnsupportedOperationException(UNMODIFIABLE);
-    }
-
-    @Override
-    public boolean addEdge(final Integer sourceVertex, final Integer targetVertex, final Integer e)
-    {
-        throw new UnsupportedOperationException(UNMODIFIABLE);
-    }
-
-    @Override
-    public Integer addVertex()
-    {
-        throw new UnsupportedOperationException(UNMODIFIABLE);
-    }
-
-    @Override
-    public boolean addVertex(final Integer v)
-    {
-        throw new UnsupportedOperationException(UNMODIFIABLE);
-    }
-
-    @Override
     public boolean containsEdge(final Integer e)
     {
         return e >= 0 && e < m;
     }
 
     @Override
-    public boolean containsVertex(final Integer v)
-    {
-        return v >= 0 && v < n;
-    }
-
-    @Override
     public Set<Integer> edgeSet()
     {
         return IntSets.fromTo(0, m);
-    }
-
-    @Override
-    public int degreeOf(final Integer vertex)
-    {
-        return inDegreeOf(vertex) + outDegreeOf(vertex);
     }
 
     @Override
@@ -346,7 +181,7 @@ public class SuccinctIntDirectedGraph
 
         for (int i = d; i-- != 0;) {
             final long source = iterator.nextLong() - base--;
-            final int e = (int) (successors.successorIndexUnsafe(source * n + t));
+            final int e = (int) (successors.successorIndexUnsafe((source << sourceShift) + t));
             assert getEdgeSource(e).longValue() == source;
             assert getEdgeTarget(e).longValue() == target;
             s.add(e);
@@ -372,96 +207,31 @@ public class SuccinctIntDirectedGraph
     }
 
     @Override
-    public Integer removeEdge(final Integer sourceVertex, final Integer targetVertex)
-    {
-        throw new UnsupportedOperationException(UNMODIFIABLE);
-    }
-
-    @Override
-    public boolean removeEdge(final Integer e)
-    {
-        throw new UnsupportedOperationException(UNMODIFIABLE);
-    }
-
-    @Override
-    public boolean removeVertex(final Integer v)
-    {
-        throw new UnsupportedOperationException(UNMODIFIABLE);
-    }
-
-    @Override
-    public Set<Integer> vertexSet()
-    {
-        return IntSets.fromTo(0, n);
-    }
-
-    @Override
     public Integer getEdgeSource(final Integer e)
     {
         assertEdgeExist(e);
-        return (int) (successors.getLong(e) / n);
+        return (int) (successors.getLong(e) >>> sourceShift);
     }
 
     @Override
     public Integer getEdgeTarget(final Integer e)
     {
         assertEdgeExist(e);
-        return (int) (successors.getLong(e) % n);
-    }
-
-    @Override
-    public GraphType getType()
-    {
-        return new DefaultGraphType.Builder()
-            .directed().weighted(false).modifiable(false).allowMultipleEdges(false)
-            .allowSelfLoops(true).build();
-    }
-
-    @Override
-    public double getEdgeWeight(final Integer e)
-    {
-        return 1.0;
-    }
-
-    @Override
-    public void setEdgeWeight(final Integer e, final double weight)
-    {
-        throw new UnsupportedOperationException(UNMODIFIABLE);
+        return (int) (successors.getLong(e) & targetMask);
     }
 
     @Override
     public Integer getEdge(final Integer sourceVertex, final Integer targetVertex)
     {
-        final long index = successors.indexOfUnsafe(sourceVertex * (long) n + targetVertex);
+        final long index =
+            successors.indexOfUnsafe(((long) sourceVertex << sourceShift) + targetVertex);
         return index != -1 ? (int) index : null;
     }
 
     @Override
     public boolean containsEdge(final Integer sourceVertex, final Integer targetVertex)
     {
-        return successors.indexOfUnsafe(sourceVertex * (long) n + targetVertex) != -1;
-    }
-
-    @Override
-    public Set<Integer> getAllEdges(final Integer sourceVertex, final Integer targetVertex)
-    {
-        final Integer edge = getEdge(sourceVertex, targetVertex);
-        return edge == null ? IntSets.EMPTY_SET : IntSets.singleton(edge);
-    }
-
-    /**
-     * Ensures that the specified vertex exists in this graph, or else throws exception.
-     *
-     * @param v vertex
-     * @return <code>true</code> if this assertion holds.
-     * @throws IllegalArgumentException if specified vertex does not exist in this graph.
-     */
-    @Override
-    protected boolean assertVertexExist(final Integer v)
-    {
-        if (v < 0 || v >= n)
-            throw new IllegalArgumentException();
-        return true;
+        return successors.indexOfUnsafe(((long) sourceVertex << sourceShift) + targetVertex) != -1;
     }
 
     /**
@@ -529,6 +299,7 @@ public class SuccinctIntDirectedGraph
             final int d = (int) (result[1] - result[0]);
             final EliasFanoIndexedMonotoneLongBigList successors = graph.successors;
             final LongBigListIterator iterator = graph.predecessors.listIterator(result[0]);
+            final int sourceShift = graph.sourceShift;
 
             return () -> new IntIterator()
             {
@@ -545,7 +316,7 @@ public class SuccinctIntDirectedGraph
                         final long source = iterator.nextLong() - base--;
                         if (skipLoops && source == target && i-- != 0)
                             return false;
-                        final long v = source * n + target;
+                        final long v = (source << sourceShift) + target;
                         assert v == successors.successor(v) : v + " != " + successors.successor(v);
                         edge = (int) successors.successorIndexUnsafe(v);
                         assert graph.getEdgeSource(edge).longValue() == source;
