+++ jgrapht-unimi-dsi/src/main/java/org/jgrapht/sux4j/SuccinctIntUndirectedGraph.java
@@ -19,31 +19,22 @@
 package org.jgrapht.sux4j;
 
 import java.io.Serializable;
-import java.util.Arrays;
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.Set;
-import java.util.function.Function;
-import java.util.function.Supplier;
 
 import org.jgrapht.Graph;
 import org.jgrapht.GraphIterables;
-import org.jgrapht.GraphType;
-import org.jgrapht.Graphs;
 import org.jgrapht.alg.util.Pair;
-import org.jgrapht.graph.AbstractGraph;
-import org.jgrapht.graph.DefaultGraphType;
 import org.jgrapht.opt.graph.sparse.SparseIntUndirectedGraph;
 
 import com.google.common.collect.Iterables;
 
-import it.unimi.dsi.fastutil.ints.IntArrays;
 import it.unimi.dsi.fastutil.ints.IntIterator;
 import it.unimi.dsi.fastutil.ints.IntOpenHashSet;
 import it.unimi.dsi.fastutil.ints.IntSet;
 import it.unimi.dsi.fastutil.ints.IntSets;
 import it.unimi.dsi.fastutil.longs.LongBigListIterator;
-import it.unimi.dsi.fastutil.longs.LongIterator;
 import it.unimi.dsi.sux4j.util.EliasFanoIndexedMonotoneLongBigList;
 import it.unimi.dsi.sux4j.util.EliasFanoMonotoneLongBigList;
 
@@ -59,151 +50,12 @@ import it.unimi.dsi.sux4j.util.EliasFanoMonotoneLongBigList;
 
 public class SuccinctIntUndirectedGraph
     extends
-    AbstractGraph<Integer, Integer>
+    AbstractSuccinctUndirectedGraph<Integer>
     implements
     Serializable
 {
     private static final long serialVersionUID = 0L;
-    protected static final String UNMODIFIABLE = "this graph is unmodifiable";
 
-    /**
-     * Turns all lists of adjacent nodes into a single monotone sequence by representing the edge
-     * <var>x</var>&nbsp;&mdash;&nbsp;<var>y</var> as <var>x</var><var>n</var> + <var>y</var>.
-     * Depending on the value of {@code sorted}, only edges with source less than or equal to the
-     * target (or vice versa) are included.
-     *
-     * @param <E> the graph edge type
-     */
-    private final static class CumulativeSuccessors<E>
-        implements
-        LongIterator
-    {
-        private final Graph<Integer, E> graph;
-        private final long n;
-        private final Function<Integer, Iterable<E>> succ;
-        private final boolean sorted;
-
-        private int x = -1, d, i, e;
-        private long next = -1;
-        private int[] s = IntArrays.EMPTY_ARRAY;
-
-        public CumulativeSuccessors(
-            final Graph<Integer, E> graph, final boolean sorted,
-            final Function<Integer, Iterable<E>> succ)
-        {
-            this.n = (int) graph.iterables().vertexCount();
-            this.graph = graph;
-            this.sorted = sorted;
-            this.succ = succ;
-        }
-
-        @Override
-        public boolean hasNext()
-        {
-            if (next != -1)
-                return true;
-            if (x == n)
-                return false;
-            while (i == d) {
-                if (++x == n)
-                    return false;
-                int d = 0;
-                for (final E e : succ.apply(x)) {
-                    final int y = Graphs.getOppositeVertex(graph, e, x);
-                    if (sorted) {
-                        if (x <= y) {
-                            s = IntArrays.grow(s, d + 1);
-                            s[d++] = y;
-                        }
-                    } else {
-                        if (x >= y) {
-                            s = IntArrays.grow(s, d + 1);
-                            s[d++] = y;
-                        }
-                    }
-                }
-                Arrays.sort(s, 0, d);
-                this.d = d;
-                i = 0;
-            }
-            // The predecessor list will not be indexed, so we can gain a few bits of space by
-            // subtracting the edge position in the list
-            next = s[i] + x * n - (sorted ? 0 : e++);
-            i++;
-            return true;
-        }
-
-        @Override
-        public long nextLong()
-        {
-            if (!hasNext())
-                throw new NoSuchElementException();
-            final long result = next;
-            next = -1;
-            return result;
-        }
-    }
-
-    /**
-     * Iterates over the cumulative degrees (starts with a zero). Depending on the value of
-     * {@code sorted}, only edges with source less than or equal to the target (or vice versa) are
-     * included.
-     *
-     * @param <E> the graph edge type
-     */
-    private final static class CumulativeDegrees<E>
-        implements
-        LongIterator
-    {
-        private final int n;
-        private int x = -1;
-        private long cumul = 0;
-        private final Function<Integer, Iterable<E>> succ;
-        private final boolean sorted;
-        private final Graph<Integer, E> graph;
-
-        public CumulativeDegrees(
-            final Graph<Integer, E> graph, final boolean sorted,
-            final Function<Integer, Iterable<E>> succ)
-        {
-            this.n = (int) graph.iterables().vertexCount();
-            this.graph = graph;
-            this.succ = succ;
-            this.sorted = sorted;
-        }
-
-        @Override
-        public boolean hasNext()
-        {
-            return x < n;
-        }
-
-        @Override
-        public long nextLong()
-        {
-            if (!hasNext())
-                throw new NoSuchElementException();
-            if (x == -1)
-                return ++x;
-            int d = 0;
-            if (sorted) {
-                for (final E e : succ.apply(x))
-                    if (x <= Graphs.getOppositeVertex(graph, e, x))
-                        d++;
-            } else {
-                for (final E e : succ.apply(x))
-                    if (x >= Graphs.getOppositeVertex(graph, e, x))
-                        d++;
-            }
-            x++;
-            return cumul += d;
-        }
-    }
-
-    /** The number of vertices in the graph. */
-    private final int n;
-    /** The number of edges in the graph. */
-    private final int m;
     /** The cumulative list of outdegrees (number of edges in sorted order, including loops). */
     private final EliasFanoIndexedMonotoneLongBigList cumulativeOutdegrees;
     /** The cumulative list of indegrees (number of edges in reversed order, including loops). */
@@ -222,6 +74,8 @@ public class SuccinctIntUndirectedGraph
      */
     public <E> SuccinctIntUndirectedGraph(final Graph<Integer, E> graph)
     {
+        super((int) graph.iterables().vertexCount(), (int) graph.iterables().edgeCount());
+
         if (graph.getType().isDirected())
             throw new IllegalArgumentException("This class supports directed graphs only");
         assert graph.getType().isUndirected();
@@ -235,9 +89,6 @@ public class SuccinctIntUndirectedGraph
                 "The number of edges (" + iterables.edgeCount() + ") is greater than "
                     + Integer.MAX_VALUE);
 
-        n = (int) iterables.vertexCount();
-        m = (int) iterables.edgeCount();
-
         cumulativeOutdegrees = new EliasFanoIndexedMonotoneLongBigList(
             n + 1, m, new CumulativeDegrees<>(graph, true, iterables::edgesOf));
         cumulativeIndegrees = new EliasFanoMonotoneLongBigList(
@@ -246,7 +97,8 @@ public class SuccinctIntUndirectedGraph
         assert cumulativeIndegrees.getLong(cumulativeIndegrees.size64() - 1) == m;
 
         successors = new EliasFanoIndexedMonotoneLongBigList(
-            m, (long) n * n, new CumulativeSuccessors<>(graph, true, iterables::outgoingEdgesOf));
+            m, (long) n << sourceShift,
+            new CumulativeSuccessors<>(graph, true, iterables::outgoingEdgesOf));
         predecessors = new EliasFanoIndexedMonotoneLongBigList(
             m, (long) n * n - m,
             new CumulativeSuccessors<>(graph, false, iterables::incomingEdgesOf));
@@ -271,51 +123,9 @@ public class SuccinctIntUndirectedGraph
     }
 
     @Override
-    public Supplier<Integer> getVertexSupplier()
-    {
-        return null;
-    }
-
-    @Override
-    public Supplier<Integer> getEdgeSupplier()
-    {
-        return null;
-    }
-
-    @Override
-    public Integer addEdge(final Integer sourceVertex, final Integer targetVertex)
-    {
-        throw new UnsupportedOperationException(UNMODIFIABLE);
-    }
-
-    @Override
-    public boolean addEdge(final Integer sourceVertex, final Integer targetVertex, final Integer e)
-    {
-        throw new UnsupportedOperationException(UNMODIFIABLE);
-    }
-
-    @Override
-    public Integer addVertex()
-    {
-        throw new UnsupportedOperationException(UNMODIFIABLE);
-    }
-
-    @Override
-    public boolean addVertex(final Integer v)
-    {
-        throw new UnsupportedOperationException(UNMODIFIABLE);
-    }
-
-    @Override
     public boolean containsEdge(final Integer e)
     {
         return e >= 0 && e < m;
-    }
-
-    @Override
-    public boolean containsVertex(final Integer v)
-    {
-        return v >= 0 && v < n;
     }
 
     @Override
@@ -343,21 +153,9 @@ public class SuccinctIntUndirectedGraph
     }
 
     @Override
-    public int inDegreeOf(final Integer vertex)
-    {
-        return degreeOf(vertex);
-    }
-
-    @Override
     public IntSet incomingEdgesOf(final Integer vertex)
     {
         return edgesOf(vertex);
-    }
-
-    @Override
-    public int outDegreeOf(final Integer vertex)
-    {
-        return degreeOf(vertex);
     }
 
     @Override
@@ -367,62 +165,17 @@ public class SuccinctIntUndirectedGraph
     }
 
     @Override
-    public Integer removeEdge(final Integer sourceVertex, final Integer targetVertex)
-    {
-        throw new UnsupportedOperationException(UNMODIFIABLE);
-    }
-
-    @Override
-    public boolean removeEdge(final Integer e)
-    {
-        throw new UnsupportedOperationException(UNMODIFIABLE);
-    }
-
-    @Override
-    public boolean removeVertex(final Integer v)
-    {
-        throw new UnsupportedOperationException(UNMODIFIABLE);
-    }
-
-    @Override
-    public Set<Integer> vertexSet()
-    {
-        return IntSets.fromTo(0, n);
-    }
-
-    @Override
     public Integer getEdgeSource(final Integer e)
     {
         assertEdgeExist(e);
-        return (int) (successors.getLong(e) / n);
+        return (int) (successors.getLong(e) >>> sourceShift);
     }
 
     @Override
     public Integer getEdgeTarget(final Integer e)
     {
         assertEdgeExist(e);
-        final long cumul = cumulativeOutdegrees.weakPredecessorUnsafe(e);
-        return (int) (successors.getLong(e) % n);
-    }
-
-    @Override
-    public GraphType getType()
-    {
-        return new DefaultGraphType.Builder()
-            .directed().weighted(false).modifiable(false).allowMultipleEdges(false)
-            .allowSelfLoops(true).build();
-    }
-
-    @Override
-    public double getEdgeWeight(final Integer e)
-    {
-        return 1.0;
-    }
-
-    @Override
-    public void setEdgeWeight(final Integer e, final double weight)
-    {
-        throw new UnsupportedOperationException(UNMODIFIABLE);
+        return (int) (successors.getLong(e) & targetMask);
     }
 
     @Override
@@ -435,43 +188,14 @@ public class SuccinctIntUndirectedGraph
             x = y;
             y = t;
         }
-        final long index = successors.indexOfUnsafe(x * (long) n + y);
+        final long index = successors.indexOfUnsafe(((long) x << sourceShift) + y);
         return index != -1 ? (int) index : null;
     }
 
     @Override
     public boolean containsEdge(final Integer sourceVertex, final Integer targetVertex)
     {
-        int x = sourceVertex;
-        int y = targetVertex;
-        if (x > y) {
-            final int t = x;
-            x = y;
-            y = t;
-        }
-        return successors.indexOfUnsafe(x * (long) n + y) != -1;
-    }
-
-    @Override
-    public Set<Integer> getAllEdges(final Integer sourceVertex, final Integer targetVertex)
-    {
-        final Integer edge = getEdge(sourceVertex, targetVertex);
-        return edge == null ? IntSets.EMPTY_SET : IntSets.singleton(edge);
-    }
-
-    /**
-     * Ensures that the specified vertex exists in this graph, or else throws exception.
-     *
-     * @param v vertex
-     * @return <code>true</code> if this assertion holds.
-     * @throws IllegalArgumentException if specified vertex does not exist in this graph.
-     */
-    @Override
-    protected boolean assertVertexExist(final Integer v)
-    {
-        if (v < 0 || v >= n)
-            throw new IllegalArgumentException();
-        return true;
+        return containsEdge(successors, sourceVertex, targetVertex);
     }
 
     /**
@@ -538,10 +262,10 @@ public class SuccinctIntUndirectedGraph
 
         private Iterable<Integer> reverseSortedEdgesOfNoLoops(final int target)
         {
-            final SuccinctIntUndirectedGraph graph = this.graph;
             final long[] result = new long[2];
             graph.cumulativeIndegrees.get(target, result);
             final int d = (int) (result[1] - result[0]);
+            final int sourceShift = graph.sourceShift;
             final EliasFanoIndexedMonotoneLongBigList successors = graph.successors;
             final LongBigListIterator iterator = graph.predecessors.listIterator(result[0]);
 
@@ -560,7 +284,7 @@ public class SuccinctIntUndirectedGraph
                         final long source = iterator.nextLong() - base--;
                         if (source == target && i-- == 0)
                             return false;
-                        final long v = source * n + target;
+                        final long v = (source << sourceShift) + target;
                         assert v == successors.successor(v) : v + " != " + successors.successor(v);
                         edge = (int) successors.successorIndexUnsafe(v);
                         assert graph.getEdgeSource(edge).longValue() == source;
