+++ jgrapht-unimi-dsi/src/main/java/org/jgrapht/sux4j/SuccinctUndirectedGraph.java
@@ -19,32 +19,22 @@
 package org.jgrapht.sux4j;
 
 import java.io.Serializable;
-import java.util.Arrays;
 import java.util.Iterator;
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.Set;
-import java.util.function.Function;
-import java.util.function.Supplier;
 
 import org.jgrapht.Graph;
 import org.jgrapht.GraphIterables;
-import org.jgrapht.GraphType;
-import org.jgrapht.Graphs;
 import org.jgrapht.alg.util.Pair;
-import org.jgrapht.graph.AbstractGraph;
-import org.jgrapht.graph.DefaultGraphType;
 import org.jgrapht.opt.graph.sparse.SparseIntUndirectedGraph;
 
 import com.google.common.collect.Iterables;
 
-import it.unimi.dsi.fastutil.ints.IntArrays;
+import it.unimi.dsi.fastutil.ints.IntIntPair;
 import it.unimi.dsi.fastutil.ints.IntIntSortedPair;
-import it.unimi.dsi.fastutil.ints.IntSets;
 import it.unimi.dsi.fastutil.longs.LongBigListIterator;
-import it.unimi.dsi.fastutil.longs.LongIterator;
 import it.unimi.dsi.fastutil.objects.ObjectOpenHashSet;
-import it.unimi.dsi.fastutil.objects.ObjectSets;
 import it.unimi.dsi.sux4j.util.EliasFanoIndexedMonotoneLongBigList;
 import it.unimi.dsi.sux4j.util.EliasFanoIndexedMonotoneLongBigList.EliasFanoIndexedMonotoneLongBigListIterator;
 import it.unimi.dsi.sux4j.util.EliasFanoMonotoneLongBigList;
@@ -61,151 +51,12 @@ import it.unimi.dsi.sux4j.util.EliasFanoMonotoneLongBigList;
 
 public class SuccinctUndirectedGraph
     extends
-    AbstractGraph<Integer, IntIntSortedPair>
+    AbstractSuccinctUndirectedGraph<IntIntSortedPair>
     implements
     Serializable
 {
     private static final long serialVersionUID = 0L;
-    protected static final String UNMODIFIABLE = "this graph is unmodifiable";
 
-    /**
-     * Turns all lists of adjacent nodes into a single monotone sequence by representing the edge
-     * <var>x</var>&nbsp;&mdash;&nbsp;<var>y</var> as <var>x</var><var>n</var> + <var>y</var>.
-     * Depending on the value of {@code sorted}, only edges with source less than or equal to the
-     * target (or vice versa) are included.
-     *
-     * @param <E> the graph edge type
-     */
-    private final static class CumulativeSuccessors<E>
-        implements
-        LongIterator
-    {
-        private final Graph<Integer, E> graph;
-        private final long n;
-        private final Function<Integer, Iterable<E>> succ;
-        private final boolean sorted;
-
-        private int x = -1, d, i, e;
-        private long next = -1;
-        private int[] s = IntArrays.EMPTY_ARRAY;
-
-        public CumulativeSuccessors(
-            final Graph<Integer, E> graph, final boolean sorted,
-            final Function<Integer, Iterable<E>> succ)
-        {
-            this.n = (int) graph.iterables().vertexCount();
-            this.graph = graph;
-            this.sorted = sorted;
-            this.succ = succ;
-        }
-
-        @Override
-        public boolean hasNext()
-        {
-            if (next != -1)
-                return true;
-            if (x == n)
-                return false;
-            while (i == d) {
-                if (++x == n)
-                    return false;
-                int d = 0;
-                for (final E e : succ.apply(x)) {
-                    final int y = Graphs.getOppositeVertex(graph, e, x);
-                    if (sorted) {
-                        if (x <= y) {
-                            s = IntArrays.grow(s, d + 1);
-                            s[d++] = y;
-                        }
-                    } else {
-                        if (x >= y) {
-                            s = IntArrays.grow(s, d + 1);
-                            s[d++] = y;
-                        }
-                    }
-                }
-                Arrays.sort(s, 0, d);
-                this.d = d;
-                i = 0;
-            }
-            // The predecessor list will not be indexed, so we can gain a few bits of space by
-            // subtracting the edge position in the list
-            next = s[i] + x * n - (sorted ? 0 : e++);
-            i++;
-            return true;
-        }
-
-        @Override
-        public long nextLong()
-        {
-            if (!hasNext())
-                throw new NoSuchElementException();
-            final long result = next;
-            next = -1;
-            return result;
-        }
-    }
-
-    /**
-     * Iterates over the cumulative degrees (starts with a zero). Depending on the value of
-     * {@code sorted}, only edges with source less than or equal to the target (or vice versa) are
-     * included.
-     *
-     * @param <E> the graph edge type
-     */
-    private final static class CumulativeDegrees<E>
-        implements
-        LongIterator
-    {
-        private final int n;
-        private int x = -1;
-        private long cumul = 0;
-        private final Function<Integer, Iterable<E>> succ;
-        private final boolean sorted;
-        private final Graph<Integer, E> graph;
-
-        public CumulativeDegrees(
-            final Graph<Integer, E> graph, final boolean sorted,
-            final Function<Integer, Iterable<E>> succ)
-        {
-            this.n = (int) graph.iterables().vertexCount();
-            this.graph = graph;
-            this.succ = succ;
-            this.sorted = sorted;
-        }
-
-        @Override
-        public boolean hasNext()
-        {
-            return x < n;
-        }
-
-        @Override
-        public long nextLong()
-        {
-            if (!hasNext())
-                throw new NoSuchElementException();
-            if (x == -1)
-                return ++x;
-            int d = 0;
-            if (sorted) {
-                for (final E e : succ.apply(x))
-                    if (x <= Graphs.getOppositeVertex(graph, e, x))
-                        d++;
-            } else {
-                for (final E e : succ.apply(x))
-                    if (x >= Graphs.getOppositeVertex(graph, e, x))
-                        d++;
-            }
-            x++;
-            return cumul += d;
-        }
-    }
-
-    /** The number of vertices in the graph. */
-    private final int n;
-    /** The number of edges in the graph. */
-    private final int m;
     /** The cumulative list of outdegrees (number of edges in sorted order, including loops). */
     private final EliasFanoIndexedMonotoneLongBigList cumulativeOutdegrees;
     /** The cumulative list of indegrees (number of edges in reversed order, including loops). */
@@ -224,6 +75,8 @@ public class SuccinctUndirectedGraph
      */
     public <E> SuccinctUndirectedGraph(final Graph<Integer, E> graph)
     {
+        super((int) graph.iterables().vertexCount(), (int) graph.iterables().edgeCount());
+
         if (graph.getType().isDirected())
             throw new IllegalArgumentException("This class supports directed graphs only");
         assert graph.getType().isUndirected();
@@ -237,9 +90,6 @@ public class SuccinctUndirectedGraph
                 "The number of edges (" + iterables.edgeCount() + ") is greater than "
                     + Integer.MAX_VALUE);
 
-        n = (int) iterables.vertexCount();
-        m = (int) iterables.edgeCount();
-
         cumulativeOutdegrees = new EliasFanoIndexedMonotoneLongBigList(
             n + 1, m, new CumulativeDegrees<>(graph, true, iterables::edgesOf));
         cumulativeIndegrees = new EliasFanoMonotoneLongBigList(
@@ -248,7 +98,8 @@ public class SuccinctUndirectedGraph
         assert cumulativeIndegrees.getLong(cumulativeIndegrees.size64() - 1) == m;
 
         successors = new EliasFanoIndexedMonotoneLongBigList(
-            m, (long) n * n, new CumulativeSuccessors<>(graph, true, iterables::outgoingEdgesOf));
+            m, (long) n << sourceShift,
+            new CumulativeSuccessors<>(graph, true, iterables::outgoingEdgesOf));
         predecessors = new EliasFanoIndexedMonotoneLongBigList(
             m, (long) n * n - m,
             new CumulativeSuccessors<>(graph, false, iterables::incomingEdgesOf));
@@ -273,52 +124,9 @@ public class SuccinctUndirectedGraph
     }
 
     @Override
-    public Supplier<Integer> getVertexSupplier()
-    {
-        return null;
-    }
-
-    @Override
-    public Supplier<IntIntSortedPair> getEdgeSupplier()
-    {
-        return null;
-    }
-
-    @Override
-    public IntIntSortedPair addEdge(final Integer sourceVertex, final Integer targetVertex)
-    {
-        throw new UnsupportedOperationException(UNMODIFIABLE);
-    }
-
-    @Override
-    public boolean addEdge(
-        final Integer sourceVertex, final Integer targetVertex, final IntIntSortedPair e)
-    {
-        throw new UnsupportedOperationException(UNMODIFIABLE);
-    }
-
-    @Override
-    public Integer addVertex()
-    {
-        throw new UnsupportedOperationException(UNMODIFIABLE);
-    }
-
-    @Override
-    public boolean addVertex(final Integer v)
-    {
-        throw new UnsupportedOperationException(UNMODIFIABLE);
-    }
-
-    @Override
     public boolean containsEdge(final IntIntSortedPair e)
     {
-        return successors.indexOfUnsafe(e.firstInt() * (long) n + e.secondInt()) != -1;
-    }
-
-    @Override
-    public boolean containsVertex(final Integer v)
-    {
-        return v >= 0 && v < n;
+        return successors.indexOfUnsafe(((long) e.firstInt() << sourceShift) + e.secondInt()) != -1;
     }
 
     @Override
@@ -344,7 +152,7 @@ public class SuccinctUndirectedGraph
 
         for (int e = (int) result[0]; e < (int) result[1]; e++) {
             final long t = successors.getLong(e);
-            s.add(IntIntSortedPair.of((int) (t / n), (int) (t % n)));
+            s.add(IntIntSortedPair.of((int) (t >>> sourceShift), (int) (t & targetMask)));
         }
 
         for (final IntIntSortedPair e : ITERABLES.reverseSortedEdgesOfNoLoops(vertex))
@@ -354,51 +162,15 @@ public class SuccinctUndirectedGraph
     }
 
     @Override
-    public int inDegreeOf(final Integer vertex)
-    {
-        return degreeOf(vertex);
-    }
-
-    @Override
     public Set<IntIntSortedPair> incomingEdgesOf(final Integer vertex)
     {
         return edgesOf(vertex);
     }
 
     @Override
-    public int outDegreeOf(final Integer vertex)
-    {
-        return degreeOf(vertex);
-    }
-
-    @Override
     public Set<IntIntSortedPair> outgoingEdgesOf(final Integer vertex)
     {
         return edgesOf(vertex);
-    }
-
-    @Override
-    public IntIntSortedPair removeEdge(final Integer sourceVertex, final Integer targetVertex)
-    {
-        throw new UnsupportedOperationException(UNMODIFIABLE);
-    }
-
-    @Override
-    public boolean removeEdge(final IntIntSortedPair e)
-    {
-        throw new UnsupportedOperationException(UNMODIFIABLE);
-    }
-
-    @Override
-    public boolean removeVertex(final Integer v)
-    {
-        throw new UnsupportedOperationException(UNMODIFIABLE);
-    }
-
-    @Override
-    public Set<Integer> vertexSet()
-    {
-        return IntSets.fromTo(0, n);
     }
 
     @Override
@@ -413,24 +185,35 @@ public class SuccinctUndirectedGraph
         return e.secondInt();
     }
 
-    @Override
-    public GraphType getType()
+    /**
+     * Returns the index associated with the given edge.
+     *
+     * @param e an edge of the graph.
+     * @return the index associated with the edge, or &minus;1 if the edge is not part of the graph.
+     * @see #getEdgeFromIndex(int)
+     */
+    public int getIndexFromEdge(final IntIntSortedPair e)
     {
-        return new DefaultGraphType.Builder()
-            .directed().weighted(false).modifiable(false).allowMultipleEdges(false)
-            .allowSelfLoops(true).build();
+        final int source = e.firstInt();
+        final int target = e.secondInt();
+        if (source < 0 || source >= n || target < 0 || target >= n)
+            throw new IllegalArgumentException();
+        return (int) successors.indexOfUnsafe(((long) source << sourceShift) + target);
     }
 
-    @Override
-    public double getEdgeWeight(final IntIntSortedPair e)
+    /**
+     * Returns the edge with given index.
+     *
+     * @param i an index between 0 (included) and the number of edges (excluded).
+     * @return the pair with index {@code i}.
+     * @see #getIndexFromEdge(IntIntPair)
+     */
+    public IntIntSortedPair getEdgeFromIndex(final int i)
     {
-        return 1.0;
-    }
-
-    @Override
-    public void setEdgeWeight(final IntIntSortedPair e, final double weight)
-    {
-        throw new UnsupportedOperationException(UNMODIFIABLE);
+        if (i < 0 || i >= m)
+            throw new IllegalArgumentException();
+        final long t = successors.getLong(i);
+        return IntIntSortedPair.of((int) (t >>> sourceShift), (int) (t & targetMask));
     }
 
     @Override
@@ -443,58 +226,14 @@ public class SuccinctUndirectedGraph
             x = y;
             y = t;
         }
-        final long index = successors.indexOfUnsafe(x * (long) n + y);
+        final long index = successors.indexOfUnsafe(((long) x << sourceShift) + y);
         return index != -1 ? IntIntSortedPair.of(x, y) : null;
     }
 
     @Override
     public boolean containsEdge(final Integer sourceVertex, final Integer targetVertex)
     {
-        int x = sourceVertex;
-        int y = targetVertex;
-        if (x > y) {
-            final int t = x;
-            x = y;
-            y = t;
-        }
-        return successors.indexOfUnsafe(x * (long) n + y) != -1;
-    }
-
-    @Override
-    public Set<IntIntSortedPair> getAllEdges(final Integer sourceVertex, final Integer targetVertex)
-    {
-        final IntIntSortedPair edge = getEdge(sourceVertex, targetVertex);
-        return edge == null ? ObjectSets.emptySet() : ObjectSets.singleton(edge);
-    }
-
-    /**
-     * Ensures that the specified vertex exists in this graph, or else throws exception.
-     *
-     * @param v vertex
-     * @return <code>true</code> if this assertion holds.
-     * @throws IllegalArgumentException if specified vertex does not exist in this graph.
-     */
-    @Override
-    protected boolean assertVertexExist(final Integer v)
-    {
-        if (v < 0 || v >= n)
-            throw new IllegalArgumentException();
-        return true;
-    }
-
-    /**
-     * Ensures that the specified edge exists in this graph, or else throws exception.
-     *
-     * @param e edge
-     * @return <code>true</code> if this assertion holds.
-     * @throws IllegalArgumentException if specified edge does not exist in this graph.
-     */
-    protected boolean assertEdgeExist(final Integer e)
-    {
-        if (e < 0 || e >= m)
-            throw new IllegalArgumentException();
-        return true;
-
+        return containsEdge(successors, sourceVertex, targetVertex);
     }
 
     private final static class SuccinctGraphIterables
@@ -536,6 +275,9 @@ public class SuccinctUndirectedGraph
         @Override
         public Iterable<IntIntSortedPair> edges()
         {
+            final int sourceShift = graph.sourceShift;
+            final long targetMask = graph.targetMask;
+
             return () -> new Iterator<>()
             {
                 private final EliasFanoIndexedMonotoneLongBigListIterator iterator =
@@ -552,7 +294,7 @@ public class SuccinctUndirectedGraph
                 public IntIntSortedPair next()
                 {
                     final long t = iterator.nextLong();
-                    return IntIntSortedPair.of((int) (t / n), (int) (t % n));
+                    return IntIntSortedPair.of((int) (t >>> sourceShift), (int) (t & targetMask));
                 }
 
             };
@@ -566,11 +308,14 @@ public class SuccinctUndirectedGraph
 
         private Iterable<IntIntSortedPair> sortedEdges(final int source)
         {
+            final int sourceShift = graph.sourceShift;
+            final long targetMask = graph.targetMask;
             final long[] result = new long[2];
             graph.cumulativeOutdegrees.get(source, result);
             final var successors = graph.successors;
             final int n = graph.n;
-            final Iterable<IntIntSortedPair> sortedEdgesIterable = () -> new Iterator<>()
+
+            return () -> new Iterator<>()
             {
                 private int e = (int) result[0];
 
@@ -584,10 +329,9 @@ public class SuccinctUndirectedGraph
                 public IntIntSortedPair next()
                 {
                     final long t = successors.getLong(e++);
-                    return IntIntSortedPair.of((int) (t / n), (int) (t % n));
+                    return IntIntSortedPair.of((int) (t >>> sourceShift), (int) (t & targetMask));
                 }
             };
-            return sortedEdgesIterable;
         }
 
         private Iterable<IntIntSortedPair> reverseSortedEdgesOfNoLoops(final int target)
